+++ src/obc/providers/base.py
@@ -3,15 +3,24 @@
 from abc import ABC, abstractmethod
 
 
-class BaseProvider(ABC):
-    """Base badge provider class."""
-
-    code: str = "BPC"
-    name: str = "Base provider"
+class BaseAssertion(ABC):
+    """Base assertion class."""
 
     @abstractmethod
-    def __init__(self, *args, **kwargs):
-        """Initialize the API client."""
+    def __init__(self, api_client):
+        """Initialize the assertion class."""
+
+    @abstractmethod
+    def read(self, assertion, query=None):
+        """Read an assertion."""
+
+
+class BaseBadge(ABC):
+    """Base badge class."""
+
+    @abstractmethod
+    def __init__(self, api_client):
+        """Initialize the badge class."""
 
     @abstractmethod
     def create(self, badge):
@@ -36,3 +45,14 @@ class BaseProvider(ABC):
     @abstractmethod
     def revoke(self, revokation):
         """Revoke one or more badges."""
+
+
+class BaseProvider(ABC):
+    """Base provider class."""
+
+    code: str = "BPC"
+    name: str = "Base provider"
+
+    @abstractmethod
+    def __init__(self, *args, **kwargs):
+        """Initialize the API client, the badge and the assertion classes."""